Switch to https, from http
Earlier versions of apt did not ship with https support, and required
that the "apt-transport-https" package be installed (which was not
default). As such, we did not enable https in our repositories
previously.

As of apt 1.6, this has changed, with the https method now being folded
into the regular apt package.

Switch to https, from http, for the protocol used to download packages.

Closes: gh-7