debian/preinst: add
Failing in postinst is awful, as it brings the system to a state that is
very difficult to recover from. Use a preinst script instead, which does
not have that problem.

However, preinst can't do certain things, such as configuring the
system, as neither our dependencies nor the contents of our own package
are guaranteed to be installed yet at the point that runs.

So copy some of the logic from postinst into preinst, but do not fail in
postinst, and do not copy everything.

Yes, this sucks :-(

While here, redo the logic using os-release rather than the (deprecated)
lsb_release script.