Generate and distribute new "continuous" key
We have been working on a replacement of the release key, 6773d225, by a
new ECDSA key, 67da21a4.

However, the "continuous" key, 10a04d46, has the same security issues
that the "official releases" one has, so we should replace that one too.

Generate the key for that, and start distributing it, but do not (yet)
use.